nishanth-submission/requirements.txt
colorama==0.4.6
Faker==26.0.0
iniconfig==2.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
py4j==0.10.9.7
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tzdata==2024.1
